Update detection_algo_2sensors.ino
Can now change type and position of each sensor, changing each associated parameters. Ready for when the protocol for determining threshold etc is finalised. However, there is one issue - the signal from sensor 1 is also picked up by sensor 2 (even when the sensors are positioned on different arms). However, the signal picked up by sensor 2 is completely separate and is not detected by sensor 1. This is solved for the desired function by line 520, only allowing the digital output 2 to be HIGH when the state of the analogue sensor is LOW. However, this is not an ideal longterm solution.